+++ phpstan.neon
@@ -33,12 +33,9 @@ parameters:
         - '#Class "Rector\\Nette\\Rector\\Class_\\CreateOrUpdateFormDataRector" is missing @see annotation with test case class reference#'
 
         - '#Cognitive complexity for "Rector\\Nette\\Kdyby\\ContributeEventClassResolver\:\:resolveGetterMethodByEventClassAndParam\(\)" is 10, keep it under 9#'
-
-        - '#Cognitive complexity for "Rector\\Nette\\NodeFinder\\FormFinder\:\:findFormVariable\(\)" is 10, keep it under 9#'
-
-        - '#Cognitive complexity for "Rector\\Nette\\NodeFinder\\FormFinder\:\:findFormFields\(\)" is 17, keep it under 9#'
-
-        - '#Cognitive complexity for "Rector\\Nette\\NodeFinder\\FormFinder\:\:findOnSuccessCallback\(\)" is 14, keep it under 9#'
+        - '#Cognitive complexity for "Rector\\Nette\\NodeFinder\\FormFieldsFinder\:\:find\(\)" is 10, keep it under 9#'
+        - '#Cognitive complexity for "Rector\\Nette\\NodeFinder\\FormOnSuccessCallbackFinder\:\:find\(\)" is 10, keep it under 9#'
+        - '#Cognitive complexity for "Rector\\Nette\\NodeFinder\\FormVariableFinder\:\:find\(\)" is 10, keep it under 9#'
 
         - '#generic class Rector\\BetterPhpDocParser\\PhpDocInfo\\PhpDocInfo but does not specify its types\: TNode#'
         - '#Use another value object over array with string\-keys and objects, array<string, ValueObject\>#'
@@ -56,15 +53,6 @@ parameters:
 
         # PHPSTan false positive
         - '#Parameter \#1 \$className (.*?) expects class\-string, string given#'
-
-        -
-            message: '#Class cognitive complexity is 61, keep it under 50#'
-            path: src/NodeFinder/FormFinder.php
-
-        # we need short class name at this place
-        -
-            message: '#Use `\$class\-\>namespaceName` instead of `\$class\-\>name` that only returns short class name#'
-            path: src/Rector/Class_/CreateOrUpdateFormDataRector.php
 
         # this is the way how it works in rector-src
         -
